Added Buld edit Mode
