+++ src/components/BindingPolicy/CreateBindingPolicyDialog.tsx
@@ -546,7 +546,7 @@ const CreateBindingPolicyDialog: React.FC<CreateBindingPolicyDialogProps> = ({
           workloadLabels[labelInfo.key] = labelInfo.value;
         }
       } else {
-        workloadLabels['kubernetes.io/kubestellar.workload.name'] = workloadObj.name;
+        workloadLabels['kubestellar.io/workload'] = workloadObj.name;
       }
       
       if (clusterId.startsWith('label-')) {
@@ -795,7 +795,7 @@ const CreateBindingPolicyDialog: React.FC<CreateBindingPolicyDialogProps> = ({
           workloadLabels[labelInfo.key] = labelInfo.value;
         }
       } else {
-        workloadLabels['kubernetes.io/kubestellar.workload.name'] = workloadObj.name;
+        workloadLabels['kubestellar.io/workload'] = workloadObj.name;
       }
       
       if (clusterId.startsWith('label-')) {
