+++ src/hooks/queries/useBPQueries.ts
@@ -636,7 +636,7 @@ export const useBPQueries = () => {
         if (!formattedRequest.workloadLabels || Object.keys(formattedRequest.workloadLabels).length === 0) {
           console.warn("No workload labels provided");
           formattedRequest.workloadLabels = {
-            'kubernetes.io/kubestellar.workload.name': 'unknown' // Fallback default
+            'kubestellar.io/workload': 'unknown' // Fallback default
           };
         }
         
@@ -703,7 +703,7 @@ export const useBPQueries = () => {
         if (!formattedRequest.workloadLabels || Object.keys(formattedRequest.workloadLabels).length === 0) {
           console.warn("No workload labels provided for YAML generation");
           formattedRequest.workloadLabels = {
-            'kubernetes.io/kubestellar.workload.name': 'unknown' // Fallback default
+            'kubestellar.io/workload': 'unknown' // Fallback default
           };
         }
